chore(base): Move the `m.room.create` wrapper types in the new `create` module.
This patch moves the `m.room.create` wrapper types, aka
`RoomCreateWithCreatorEventContent` type and siblings, in the new
`create` module.
