+++ crates/matrix-sdk-base/src/rooms/mod.rs
@@ -14,6 +14,7 @@
 
 #![allow(clippy::assign_op_pattern)] // Triggered by bitflags! usage
 
+mod create;
 mod display_name;
 mod encryption;
 mod knock;
@@ -32,6 +33,7 @@ use crate::{
     Error, MinimalStateEvent,
 };
 use as_variant::as_variant;
+pub use create::*;
 pub use display_name::{RoomDisplayName, RoomHero};
 pub(crate) use display_name::{RoomSummary, UpdatedRoomDisplayName};
 pub use encryption::EncryptionState;
@@ -44,14 +46,13 @@ pub(crate) use room_info::SyncInfo;
 pub use room_info::{
     apply_redaction, BaseRoomInfo, RoomInfo, RoomInfoNotableUpdate, RoomInfoNotableUpdateReasons,
 };
+#[cfg(feature = "e2e-encryption")]
+use ruma::{events::AnySyncTimelineEvent, serde::Raw};
 use ruma::{
-    assign,
     events::{
         direct::OwnedDirectUserIdentifier,
-        macros::EventContent,
         member_hints::MemberHintsEventContent,
         receipt::{Receipt, ReceiptThread, ReceiptType},
-        room::create::{PreviousRoom, RoomCreateEventContent},
         room::{
             avatar::{self},
             guest_access::GuestAccess,
@@ -60,14 +61,11 @@ use ruma::{
             power_levels::{RoomPowerLevels, RoomPowerLevelsEventContent},
             tombstone::RoomTombstoneEventContent,
         },
-        EmptyStateKey, RedactContent, RedactedStateEventContent, SyncStateEvent,
+        SyncStateEvent,
     },
     room::RoomType,
-    EventId, OwnedEventId, OwnedMxcUri, OwnedRoomAliasId, OwnedRoomId, OwnedUserId, RoomId,
-    RoomVersionId, UserId,
+    EventId, OwnedEventId, OwnedMxcUri, OwnedRoomAliasId, OwnedRoomId, OwnedUserId, RoomId, UserId,
 };
-#[cfg(feature = "e2e-encryption")]
-use ruma::{events::AnySyncTimelineEvent, serde::Raw};
 use serde::{Deserialize, Serialize};
 pub use state::{RoomState, RoomStateFilter};
 #[cfg(feature = "e2e-encryption")]
@@ -532,99 +530,6 @@ fn test_send_sync_for_room() {
     fn assert_send_sync<T: Send + Sync>() {}
 
     assert_send_sync::<Room>();
-}
-
-/// The content of an `m.room.create` event, with a required `creator` field.
-///
-/// Starting with room version 11, the `creator` field should be removed and the
-/// `sender` field of the event should be used instead. This is reflected on
-/// [`RoomCreateEventContent`].
-///
-/// This type was created as an alternative for ease of use. When it is used in
-/// the SDK, it is constructed by copying the `sender` of the original event as
-/// the `creator`.
-#[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
-#[ruma_event(type = "m.room.create", kind = State, state_key_type = EmptyStateKey, custom_redacted)]
-pub struct RoomCreateWithCreatorEventContent {
-    /// The `user_id` of the room creator.
-    ///
-    /// This is set by the homeserver.
-    ///
-    /// While this should be optional since room version 11, we copy the sender
-    /// of the event so we can still access it.
-    pub creator: OwnedUserId,
-
-    /// Whether or not this room's data should be transferred to other
-    /// homeservers.
-    #[serde(
-        rename = "m.federate",
-        default = "ruma::serde::default_true",
-        skip_serializing_if = "ruma::serde::is_true"
-    )]
-    pub federate: bool,
-
-    /// The version of the room.
-    ///
-    /// Defaults to `RoomVersionId::V1`.
-    #[serde(default = "default_create_room_version_id")]
-    pub room_version: RoomVersionId,
-
-    /// A reference to the room this room replaces, if the previous room was
-    /// upgraded.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub predecessor: Option<PreviousRoom>,
-
-    /// The room type.
-    ///
-    /// This is currently only used for spaces.
-    #[serde(skip_serializing_if = "Option::is_none", rename = "type")]
-    pub room_type: Option<RoomType>,
-}
-
-impl RoomCreateWithCreatorEventContent {
-    /// Constructs a `RoomCreateWithCreatorEventContent` with the given original
-    /// content and sender.
-    pub fn from_event_content(content: RoomCreateEventContent, sender: OwnedUserId) -> Self {
-        let RoomCreateEventContent { federate, room_version, predecessor, room_type, .. } = content;
-        Self { creator: sender, federate, room_version, predecessor, room_type }
-    }
-
-    fn into_event_content(self) -> (RoomCreateEventContent, OwnedUserId) {
-        let Self { creator, federate, room_version, predecessor, room_type } = self;
-
-        #[allow(deprecated)]
-        let content = assign!(RoomCreateEventContent::new_v11(), {
-            creator: Some(creator.clone()),
-            federate,
-            room_version,
-            predecessor,
-            room_type,
-        });
-
-        (content, creator)
-    }
-}
-
-/// Redacted form of [`RoomCreateWithCreatorEventContent`].
-pub type RedactedRoomCreateWithCreatorEventContent = RoomCreateWithCreatorEventContent;
-
-impl RedactedStateEventContent for RedactedRoomCreateWithCreatorEventContent {
-    type StateKey = EmptyStateKey;
-}
-
-impl RedactContent for RoomCreateWithCreatorEventContent {
-    type Redacted = RedactedRoomCreateWithCreatorEventContent;
-
-    fn redact(self, version: &RoomVersionId) -> Self::Redacted {
-        let (content, sender) = self.into_event_content();
-        // Use Ruma's redaction algorithm.
-        let content = content.redact(version);
-        Self::from_event_content(content, sender)
-    }
-}
-
-fn default_create_room_version_id() -> RoomVersionId {
-    RoomVersionId::V1
 }
 
 /// The possible sources of an account data type.
